Improve postgres db creation by specifying more args
The current db create make task wasn't using the db configuration
for anything but the db name and so the command would fail if it
required login or the db was on another host (like in docker).
These changes make it use values from the environment and then
fall back to using values from the config in order to specify
arguments to createdb.

Postgres env vars were borrowed from:
https://www.postgresql.org/docs/12/libpq-envars.html