Set LANG=c when generating secret
The character type in Monterey appears to be set to an unexpected value,
causing the `coast` script to raise an errer when attempting to generate
a random secret.

On my system, LANG was set to 'en_CA.UTF-8'. When piping the output of
/dev/urandom to `fold`, a message 'Input error' appears and an empty
string is returned and set as the "secret". ie, any place that we expect
to have a random secret set for our app will actually have an empty
string!

This change is the quickest fix I could think of to correct this.
Another option would be to set LC_CTYPE to `en.US.UTF-8`, but I'm not
sure if this is compatible on other systems.