Allow verily validations to be fully leveraged.
Currently, coast's validations only permits vector form of verily validations. This makes it impossible to directly use custom validations, including domain-specific validations; or to compose validations, such as would be possible via verily's `combine` function - especially useful when large parts of a validation are shared between creation and update endpoints.

By checking for a function as the input validations, it may be possible to permit the full scope of verily's 'custom validator fn' functionality and its composing functionality. This does not validate the shape of the fn as of yet, so further constraints may be desirable before merging.