Attempt to fix split up/down Clojure migrations.
Currently, contrary to the docs, using split up/down Clojure migrations does not run as intended; instead, it runs the *last defined function in the migration file*, which is often going to be `down` by convention.

This is an attempt to guess the intent of a migration by checking for `change` functions explicitly, running it if found, and trying to run the `up` and `down` migrations otherwise. It also fixes a potential issue with `down` migrations being inverted by code intended to invert `change` migrations automatically.

This is an ugly draft, so more would need to be done here to proof it, I expect.